requirements.txt
# Core Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2

# CORS and Security
django-cors-headers==4.3.0
django-ratelimit==4.1.0

# Password hashing (stronger than default)
argon2-cffi==23.1.0

# Filtering and Search
django-filter==23.5

# Channels for WebSocket (messaging)
channels==4.0.0
channels-redis==4.1.0

# Celery for background tasks
celery==5.3.4
redis==5.0.1

# Environment variables
python-decouple==3.8

# Image handling
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production server
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# Development
ipython==8.18.1
django-extensions==3.2.3
